# openapi/luminaa-openapi-spec.json
{"openapi":"3.0.0","paths":{"/api/v1":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/api/v1/users":{"post":{"operationId":"UsersController_create","summary":"Create user","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"200":{"description":"New User created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"403":{"description":"Forbidden."}},"tags":["users"],"security":[{"bearer":[]}]},"get":{"operationId":"UsersController_findAll","summary":"Get all users","parameters":[],"responses":{"200":{"description":"List of all Users","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}},"403":{"description":"Forbidden."}},"tags":["users"],"security":[{"bearer":[]}]}},"/api/v1/users/{id}":{"get":{"operationId":"UsersController_findOne","summary":"Get user by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The found record","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"403":{"description":"Forbidden."}},"tags":["users"],"security":[{"bearer":[]}]},"patch":{"operationId":"UsersController_update","summary":"Update User by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"Updated","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"403":{"description":"Forbidden."}},"tags":["users"],"security":[{"bearer":[]}]},"delete":{"operationId":"UsersController_remove","summary":"Delete User by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Deleted"},"403":{"description":"Forbidden."}},"tags":["users"],"security":[{"bearer":[]}]}},"/api/v1/auth/register":{"post":{"operationId":"AuthController_register","summary":"Register a user","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"200":{"description":"New User created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"403":{"description":"Forbidden."}},"tags":["auth"]}},"/api/v1/auth/login":{"post":{"operationId":"AuthController_login","summary":"Login a user","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"User was logedin","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"403":{"description":"Forbidden."}},"tags":["auth"]}},"/api/v1/auth/pin/verify":{"post":{"operationId":"AuthController_pinVerify","summary":"Check Pin of user","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PinVerifyDto"}}}},"responses":{"200":{"description":"User PIN is right","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"403":{"description":"Forbidden."}},"tags":["auth"]}},"/api/v1/auth/pin/register":{"post":{"operationId":"AuthController_pinRegister","summary":"Register Pin for user","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PinRegisterDto"}}}},"responses":{"200":{"description":"PIN is registered","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"403":{"description":"Forbidden."}},"tags":["auth"]}},"/api/v1/auth/refresh":{"get":{"operationId":"AuthController_refresh","summary":"Refresh a user token","parameters":[],"responses":{"200":{"description":"User token was refreshed"},"403":{"description":"Forbidden."}},"tags":["auth"],"security":[{"bearer":[]}]}},"/api/v1/groups/{id}/groupchats/{id}/messages":{"post":{"operationId":"MessagesController_create","summary":"Create message","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateMessageDto"}}}},"responses":{"200":{"description":"New Message created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Message"}}}},"403":{"description":"Forbidden."}},"tags":["messages"],"security":[{"bearer":[]}]},"get":{"operationId":"MessagesController_findAll","summary":"Get all messages","parameters":[{"name":"requesterId","required":true,"in":"query","schema":{"type":"string"}},{"name":"chatId","required":false,"in":"query","schema":{"type":"string"}},{"name":"senderId","required":false,"in":"query","schema":{}}],"responses":{"200":{"description":"List of all Messages","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Message"}}}}},"403":{"description":"Forbidden."}},"tags":["messages"],"security":[{"bearer":[]}]}},"/api/v1/groups/{id}/groupchats/{id}/messages/{id}":{"get":{"operationId":"MessagesController_findOne","summary":"Get Message by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The found record","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Message"}}}},"403":{"description":"Forbidden."}},"tags":["messages"],"security":[{"bearer":[]}]},"patch":{"operationId":"MessagesController_update","summary":"Update Message by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateMessageDto"}}}},"responses":{"200":{"description":"Updated","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Message"}}}},"403":{"description":"Forbidden."}},"tags":["messages"],"security":[{"bearer":[]}]},"delete":{"operationId":"MessagesController_remove","summary":"Delete Message by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Deleted"},"403":{"description":"Forbidden."}},"tags":["messages"],"security":[{"bearer":[]}]}},"/api/v1/groups":{"post":{"operationId":"GroupsController_create","summary":"Create Group","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateGroupDto"}}}},"responses":{"200":{"description":"New Group created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Group"}}}},"403":{"description":"Forbidden."}},"tags":["groups"],"security":[{"bearer":[]}]},"get":{"operationId":"GroupsController_findAll","summary":"Get all Groups","parameters":[],"responses":{"200":{"description":"List of all Groups","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Group"}}}}},"403":{"description":"Forbidden."}},"tags":["groups"],"security":[{"bearer":[]}]}},"/api/v1/groups/{id}":{"get":{"operationId":"GroupsController_findOne","summary":"Get Group by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The found record","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Group"}}}},"403":{"description":"Forbidden."}},"tags":["groups"],"security":[{"bearer":[]}]},"patch":{"operationId":"GroupsController_update","summary":"Update Group by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateGroupDto"}}}},"responses":{"200":{"description":"Updated","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Group"}}}},"403":{"description":"Forbidden."}},"tags":["groups"],"security":[{"bearer":[]}]},"delete":{"operationId":"GroupsController_remove","summary":"Delete Group by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Deleted"},"403":{"description":"Forbidden."}},"tags":["groups"],"security":[{"bearer":[]}]}},"/api/v1/roles":{"post":{"operationId":"RolesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRoleDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"RolesController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/api/v1/roles/{id}":{"get":{"operationId":"RolesController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"RolesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRoleDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"RolesController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/api/v1/groups/{id}/groupchats":{"post":{"operationId":"GroupChatsController_create","summary":"Create GroupChat","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateGroupChatDto"}}}},"responses":{"200":{"description":"New GroupChat created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GroupChat"}}}},"403":{"description":"Forbidden."}},"tags":["groupchats"],"security":[{"bearer":[]}]},"get":{"operationId":"GroupChatsController_findAll","summary":"Get all GroupChats","parameters":[],"responses":{"200":{"description":"List of all GroupChats","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GroupChat"}}}}},"403":{"description":"Forbidden."}},"tags":["groupchats"],"security":[{"bearer":[]}]}},"/api/v1/groups/{id}/groupchats/{id}":{"get":{"operationId":"GroupChatsController_findOne","summary":"Get GroupChat by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The found record","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GroupChat"}}}},"403":{"description":"Forbidden."}},"tags":["groupchats"],"security":[{"bearer":[]}]},"patch":{"operationId":"GroupChatsController_update","summary":"Update GroupChat by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateGroupChatDto"}}}},"responses":{"200":{"description":"Updated","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GroupChat"}}}},"403":{"description":"Forbidden."}},"tags":["groupchats"],"security":[{"bearer":[]}]},"delete":{"operationId":"GroupChatsController_remove","summary":"Delete GroupChat by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Deleted"},"403":{"description":"Forbidden."}},"tags":["groupchats"],"security":[{"bearer":[]}]}},"/api/v1/group-type":{"post":{"operationId":"GroupTypeController_create","summary":"Create Group Type","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateGroupTypeDto"}}}},"responses":{"200":{"description":"New Group Type created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GroupType"}}}},"403":{"description":"Forbidden."}},"tags":["group-type"],"security":[{"bearer":[]}]},"get":{"operationId":"GroupTypeController_findAll","summary":"Get all Group Types","parameters":[],"responses":{"200":{"description":"List of all Group Types","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GroupType"}}}}},"403":{"description":"Forbidden."}},"tags":["group-type"],"security":[{"bearer":[]}]}},"/api/v1/group-type/{id}":{"get":{"operationId":"GroupTypeController_findOne","summary":"Get Group Type by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The found record","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GroupType"}}}},"403":{"description":"Forbidden."}},"tags":["group-type"],"security":[{"bearer":[]}]},"patch":{"operationId":"GroupTypeController_update","summary":"Update Group Type by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateGroupTypeDto"}}}},"responses":{"200":{"description":"Updated","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GroupType"}}}},"403":{"description":"Forbidden."}},"tags":["group-type"],"security":[{"bearer":[]}]},"delete":{"operationId":"GroupTypeController_remove","summary":"Delete Group Type by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Deleted"},"403":{"description":"Forbidden."}},"tags":["group-type"],"security":[{"bearer":[]}]}}},"info":{"title":"Luminaa.Chat","description":"The luminaa.chat API description","version":"1.0","contact":{}},"tags":[],"servers":[{"url":"http://localhost:3000/"}],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateUserDto":{"type":"object","properties":{"email":{"type":"string","example":"johndoe@luminaa.chat","description":"Username of the User"},"name":{"type":"string","example":"John Doe","description":"E-Mail of the User"},"password":{"type":"string","example":"123456789","description":"Password of the User"},"lastLoginAt":{"type":"string","example":"2023-05-19T16:21:28.120Z","description":"Last Login At"},"role":{"type":"string","example":"customer","description":"Role of the User"},"emailVerified":{"type":"boolean","example":true,"description":"Is Email Verified"}},"required":["email","name","password","lastLoginAt","role","emailVerified"]},"User":{"type":"object","properties":{"_id":{"type":"string","example":"1aa21eb9-42d5-467a-9988-5696bc4bbda6","description":"ID of the User"},"__v":{"type":"number","example":23,"description":"Version of the User"},"email":{"type":"string","example":"info@luminaa.chat","description":"Username of the User"},"firstName":{"type":"string","example":"John","description":"Firstname of the User"},"lastName":{"type":"string","example":"Doe","description":"Lastname of the User"},"password":{"type":"string","description":"Password of the User"},"pin":{"type":"string","description":"PIN of the User"},"lastLoginAt":{"format":"date-time","type":"string","example":"2023-05-19T16:21:28.120Z","description":"Last Login At"},"expire":{"format":"date-time","type":"string","example":"2023-05-19T16:21:28.120Z","description":"Expire Date"},"active":{"type":"boolean","example":true,"description":"Is Active Status"},"roles":{"example":["operator","employee"],"description":"Role of the User","type":"array","items":{"type":"string"}},"emailVerified":{"type":"boolean","example":true,"description":"Is Email Verified"},"createdAt":{"format":"date-time","type":"string","example":"2023-05-19T16:21:28.120Z","description":"Created At of the Store"},"updatedAt":{"format":"date-time","type":"string","example":"2023-05-19T16:21:28.120Z","description":"Updated At of the Store"}},"required":["_id","__v","email","firstName","lastName","password","pin","lastLoginAt","expire","active","roles","emailVerified","createdAt","updatedAt"]},"UpdateUserDto":{"type":"object","properties":{}},"RegisterDto":{"type":"object","properties":{"email":{"type":"string","example":"johndoe@luminaa.chat","description":"Email Address of the User"},"password":{"type":"string","example":"123456789","description":"Passwort of the User"},"firstName":{"type":"string","example":"John","description":"Firstname of the User"},"lastName":{"type":"string","example":"Doe","description":"Lastname of the User"},"roles":{"example":["operator","employee"],"description":"Roles of the User","type":"array","items":{"type":"string"}}},"required":["email","password","firstName","lastName","roles"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string","example":"info@luminaa.chat","description":"Email Address of the User"},"password":{"type":"string","example":"123456789","description":"Passwort of the User"},"user":{"description":"User Data","allOf":[{"$ref":"#/components/schemas/User"}]}},"required":["email","password","user"]},"PinVerifyDto":{"type":"object","properties":{"pin":{"type":"string","example":"1234","description":"User PIN of the User"},"userID":{"type":"string","example":"12312-eqfrf2323-2323rdfdf","description":"ID of the User"}},"required":["pin","userID"]},"PinRegisterDto":{"type":"object","properties":{"pin":{"type":"string","example":"1234","description":"User PIN of the User"},"userID":{"type":"string","example":"12312-eqfrf2323-2323rdfdf","description":"ID of the User"}},"required":["pin","userID"]},"CreateMessageDto":{"type":"object","properties":{"message":{"type":"string","example":"Hello World!","description":"Message of the Message"},"sender":{"type":"string","description":"Sender of the Config"},"receiver":{"type":"string","description":"Receiver of the Config"},"type":{"type":"string","example":"Channel | Direct","description":"Type of the Message"}},"required":["message","sender","receiver","type"]},"Message":{"type":"object","properties":{"_id":{"type":"string","example":"1aa21eb9-42d5-467a-9988-5696bc4bbda6","description":"ID of the Message"},"__v":{"type":"number","example":23,"description":"Version of the Message"},"message":{"type":"string","example":"Hello World","description":"Message of the Message"},"sender":{"description":"Sender of the Message","allOf":[{"$ref":"#/components/schemas/User"}]},"receiver":{"description":"Receiver of the Message","allOf":[{"$ref":"#/components/schemas/User"}]},"type":{"type":"string","example":"channel | direct","description":"Type of the Message"},"deliveredAt":{"format":"date-time","type":"string","example":"2023-05-19T16:21:28.120Z","description":"Delivered At of the Message"},"seenAt":{"format":"date-time","type":"string","example":"2023-05-19T16:21:28.120Z","description":"Seen At of the Message"},"createdAt":{"format":"date-time","type":"string","example":"2023-05-19T16:21:28.120Z","description":"Created At of the Message"},"updatedAt":{"format":"date-time","type":"string","example":"2023-05-19T16:21:28.120Z","description":"Updated At of the Message"},"deletedAt":{"format":"date-time","type":"string","example":"2023-05-19T16:21:28.120Z","description":"DeleteAt of the Message"},"Visible":{"type":"boolean","description":"Message is currently visible"}},"required":["_id","__v","message","sender","receiver","type","deliveredAt","seenAt","createdAt","updatedAt","deletedAt","Visible"]},"UpdateMessageDto":{"type":"object","properties":{"message":{"type":"string","example":"Hello World!","description":"Message of the Message"},"sender":{"type":"string","description":"Sender of the Config"},"receiver":{"type":"string","description":"Receiver of the Config"},"type":{"type":"string","example":"Channel | Direct","description":"Type of the Message"}}},"CreateGroupDto":{"type":"object","properties":{"name":{"type":"string","example":"Cat Room","description":"Name of the Room"},"tag":{"type":"string","example":"cat-lovers","description":"Tag of the Room"},"description":{"type":"string","example":"This is a awesome channel for all cat lovers","description":"Description of the Room"},"owners":{"description":"Owners of the Room","type":"array","items":{"type":"string"}},"members":{"description":"Members of the Room","type":"array","items":{"type":"string"}},"roles":{"description":"Roles of the Room","type":"array","items":{"type":"string"}}},"required":["name","tag","description","owners","members","roles"]},"GroupType":{"type":"object","properties":{"_id":{"type":"string","example":"1aa21eb9-42d5-467a-9988-5696bc4bbda6","description":"ID of the GroupType"},"__v":{"type":"number","example":23,"description":"Version of the GroupType"},"name":{"type":"string","description":"Name of the GroupType"},"description":{"type":"string","description":"Description of the GroupType"},"createdAt":{"format":"date-time","type":"string","example":"2023-05-19T16:21:28.120Z","description":"Created At of the GroupType"},"updatedAt":{"format":"date-time","type":"string","example":"2023-05-19T16:21:28.120Z","description":"Updated At of the GroupType"},"color":{"type":"string","description":"Colorcode to be used in displaying the GroupType in collapsed views"},"Icon":{"type":"string","description":"Icon to be used in displaying the GroupType in collapsed views"},"Enabled":{"type":"boolean","description":"GroupType is currently enabled"}},"required":["_id","__v","name","description","createdAt","updatedAt","color","Icon","Enabled"]},"Group":{"type":"object","properties":{"_id":{"type":"string","example":"1aa21eb9-42d5-467a-9988-5696bc4bbda6","description":"ID of the Group"},"__v":{"type":"number","example":23,"description":"Version of the Group"},"name":{"type":"string","example":"Cat Channel","description":"Name of the Group"},"tag":{"type":"string","example":"cat-lovers","description":"Tag of the Group"},"description":{"type":"string","example":"This is a awesome channel for all cat lovers","description":"Description of the Group"},"owners":{"description":"Owners of the Group","type":"array","items":{"type":"string"}},"members":{"description":"Members of the Group","type":"array","items":{"type":"string"}},"roles":{"description":"Roles of the Group","type":"array","items":{"type":"string"}},"groupChats":{"description":"Associated Groupchats of the Group","type":"array","items":{"type":"string"}},"createdAt":{"format":"date-time","type":"string","example":"2023-05-19T16:21:28.120Z","description":"Created At of the Group"},"updatedAt":{"format":"date-time","type":"string","example":"2023-05-19T16:21:28.120Z","description":"Updated At of the Group"},"GroupType":{"description":"Type of the group","allOf":[{"$ref":"#/components/schemas/GroupType"}]},"color":{"type":"string","description":"Colorcode to be used in displaying the Group"},"Icon":{"type":"string","description":"Icon to be used in displaying the Group"},"Enabled":{"type":"boolean","description":"Group is currently enabled"}},"required":["_id","__v","name","tag","description","owners","members","roles","groupChats","createdAt","updatedAt","GroupType","color","Icon","Enabled"]},"UpdateGroupDto":{"type":"object","properties":{"name":{"type":"string","example":"Cat Room","description":"Name of the Room"},"tag":{"type":"string","example":"cat-lovers","description":"Tag of the Room"},"description":{"type":"string","example":"This is a awesome channel for all cat lovers","description":"Description of the Room"},"owners":{"description":"Owners of the Room","type":"array","items":{"type":"string"}},"members":{"description":"Members of the Room","type":"array","items":{"type":"string"}},"roles":{"description":"Roles of the Room","type":"array","items":{"type":"string"}}}},"CreateRoleDto":{"type":"object","properties":{}},"UpdateRoleDto":{"type":"object","properties":{}},"CreateGroupChatDto":{"type":"object","properties":{"name":{"type":"string","example":"Cat GroupChat","description":"Name of the GroupChat"},"tag":{"type":"string","example":"cat-lovers","description":"Tag of the GroupChat"},"description":{"type":"string","example":"This is a awesome channel for all cat lovers","description":"Description of the GroupChat"},"owners":{"description":"Owners of the Room","type":"array","items":{"type":"string"}},"members":{"description":"Members of the Room","type":"array","items":{"type":"string"}},"roles":{"description":"Roles of the Room","type":"array","items":{"type":"string"}}},"required":["name","tag","description","owners","members","roles"]},"GroupChat":{"type":"object","properties":{"_id":{"type":"string","example":"1aa21eb9-42d5-467a-9988-5696bc4bbda6","description":"ID of the GroupChat"},"__v":{"type":"number","example":23,"description":"Version of the GroupChat"},"name":{"type":"string","example":"Cat Channel","description":"Name of the GroupChat"},"tag":{"type":"string","example":"cat-lovers","description":"Tag of the GroupChat"},"description":{"type":"string","example":"This is a awesome channel for all cat lovers","description":"Description of the GroupChat"},"owners":{"description":"Owners of the GroupChat","type":"array","items":{"type":"string"}},"members":{"description":"Members of the GroupChat","type":"array","items":{"type":"string"}},"roles":{"description":"Roles of the GroupChat","type":"array","items":{"type":"string"}},"messages":{"description":"Messages of the GroupChat","type":"array","items":{"type":"string"}},"createdAt":{"format":"date-time","type":"string","example":"2023-05-19T16:21:28.120Z","description":"Created At of the GroupChat"},"updatedAt":{"format":"date-time","type":"string","example":"2023-05-19T16:21:28.120Z","description":"Updated At of the GroupChat"},"Enabled":{"type":"boolean","description":"GroupChat is currently enabled"}},"required":["_id","__v","name","tag","description","owners","members","roles","messages","createdAt","updatedAt","Enabled"]},"UpdateGroupChatDto":{"type":"object","properties":{"name":{"type":"string","example":"Cat GroupChat","description":"Name of the GroupChat"},"tag":{"type":"string","example":"cat-lovers","description":"Tag of the GroupChat"},"description":{"type":"string","example":"This is a awesome channel for all cat lovers","description":"Description of the GroupChat"},"owners":{"description":"Owners of the Room","type":"array","items":{"type":"string"}},"members":{"description":"Members of the Room","type":"array","items":{"type":"string"}},"roles":{"description":"Roles of the Room","type":"array","items":{"type":"string"}}}},"CreateGroupTypeDto":{"type":"object","properties":{"name":{"type":"string","example":"Ambulant","description":"Name of the GroupType"},"description":{"type":"string","example":"Dies ist der Ambulanten Bereich","description":"Description of the GroupType"}},"required":["name","description"]},"UpdateGroupTypeDto":{"type":"object","properties":{"name":{"type":"string","example":"Ambulant","description":"Name of the GroupType"},"description":{"type":"string","example":"Dies ist der Ambulanten Bereich","description":"Description of the GroupType"}}}}},"externalDocs":{"description":"Docs Luminaa.Chat","url":"https://docs.luminaa.chat"}}